finished tests for Inventory page
