+++ vite.config.ts
@@ -5,11 +5,6 @@ import react from "@vitejs/plugin-react";
 // https://vitejs.dev/config/
 export default defineConfig({
   plugins: [react()],
-  test: {
-    environment: 'jsdom',
-    globals: true,
-    setupFiles: './__tests__/Setup.ts'
-  },
   resolve: {
     alias: {
       "@": path.resolve(__dirname, "./src"),
